requirements.txt
fire
ipython
jupyter
joblib
matplotlib
notebook
numpy
ninja
open3d-python==0.7.0.0
opencv-python
pandas
Pillow
protobuf
pytorch-ignite==0.2.0
pyflann3==1.8.4.1
pytz
PyYAML
scikit-image
scipy
tqdm
torch==1.2.0
torchvision==0.4.0
tensorboard
typing
yacs
